# backend/Minigram.Api/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Minigram API",
    "description": "A minimalistic message server API that supports group chats.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/friendships": {
      "get": {
        "tags": [
          "Friendships"
        ],
        "summary": "List friends",
        "operationId": "Friendships_ListFriends",
        "parameters": [
          {
            "name": "searchText",
            "in": "query",
            "description": "The name of the friend",
            "schema": {
              "type": "string",
              "default": "",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "pageIndex",
            "in": "query",
            "description": "The index of the page",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 2
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The amount of items per page",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            },
            "x-position": 3
          }
        ],
        "responses": {
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListDtoOfFriendshipDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2": []
          }
        ]
      },
      "post": {
        "tags": [
          "Friendships"
        ],
        "summary": "Accept a friend request",
        "operationId": "Friendships_CreateFriendship",
        "requestBody": {
          "x-name": "dto",
          "description": "The details of the friendship",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendshipCreateDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2": []
          }
        ]
      }
    },
    "/api/friendships/{friendshipId}": {
      "delete": {
        "tags": [
          "Friendships"
        ],
        "summary": "Delete a friendship",
        "operationId": "Friendships_DeleteFriendship",
        "parameters": [
          {
            "name": "friendshipId",
            "in": "path",
            "required": true,
            "description": "The id of the friendship",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "OAuth2": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
      "UserListDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PagedListDtoOfFriendshipDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FriendshipDto"
            }
          }
        }
      },
      "FriendshipDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "friend": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/UserListDto"
              }
            ]
          }
        }
      },
      "FriendshipCreateDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "scheme": "Bearer",
        "openIdConnectUrl": "https://localhost:5101/.well-known/openid-configuration",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://localhost:5101/connect/authorize",
            "tokenUrl": "https://localhost:5101/connect/token",
            "scopes": {
              "openid": "OpenId",
              "Conversations.Read": "Read conversations",
              "Conversations.Manage": "Manage conversations",
              "Friendships.Read": "Read friendships and friend requests",
              "Friendships.Manage": "Manage friendships and friend requests"
            }
          }
        }
      }
    }
  }
}